djangotest/requirements.txt
django==4.2.3
mysqlclient==2.2.0
django-rest-framework==0.1.0

